add keep_html option for github_document (closes #1648) (#1650)
* add keep_html option for github_document (closes #1648)

* update NEWS

+++ R/github_document.R
@@ -12,6 +12,8 @@
 #'   newline to represent a line break (as opposed to two-spaces and a newline).
 #' @param html_preview \code{TRUE} to also generate an HTML file for the purpose of
 #'   locally previewing what the document will look like on GitHub.
+#' @param keep_html \code{TRUE} to keep the preview HTML file in the working
+#'   directory. Default is \code{FALSE}.
 #' @return R Markdown output format to pass to \code{\link{render}}
 #' @export
 github_document <- function(toc = FALSE,
@@ -24,7 +26,8 @@ github_document <- function(toc = FALSE,
                             md_extensions = NULL,
                             hard_line_breaks = TRUE,
                             pandoc_args = NULL,
-                            html_preview = TRUE) {
+                            html_preview = TRUE,
+                            keep_html = FALSE) {
 
   # add special markdown rendering template to ensure we include the title fields
   pandoc_args <- c(
@@ -75,12 +78,14 @@ github_document <- function(toc = FALSE,
       )
 
       # move the preview to the preview_dir if specified
-      preview_dir <- Sys.getenv("RMARKDOWN_PREVIEW_DIR", unset = NA)
-      if (!is.na(preview_dir)) {
-        relocated_preview_file <- tempfile("preview-", preview_dir, ".html")
-        file.copy(preview_file, relocated_preview_file)
-        file.remove(preview_file)
-        preview_file <- relocated_preview_file
+      if (!keep_html) {
+        preview_dir <- Sys.getenv("RMARKDOWN_PREVIEW_DIR", unset = NA)
+        if (!is.na(preview_dir)) {
+          relocated_preview_file <- tempfile("preview-", preview_dir, ".html")
+          file.copy(preview_file, relocated_preview_file)
+          file.remove(preview_file)
+          preview_file <- relocated_preview_file
+        }
       }
 
       if (verbose) message("\nPreview created: ", preview_file)
